speechPlayer dll's speechPlayer_queueFrame: minFrameDuration and fadeDuration are now in samples not miliseconds. However, The Python implementation remains in ms for ease of use.

+++ src/speechPlayer.h
@@ -25,7 +25,7 @@ extern "C" {
 typedef void* speechPlayer_handle_t;
 
 speechPlayer_handle_t speechPlayer_initialize(int sampleRate);
-void speechPlayer_queueFrame(speechPlayer_handle_t playerHandle, speechPlayer_frame_t* framePtr, double minFrameDuration, double fadeDuration, int userIndex, bool purgeQueue);
+void speechPlayer_queueFrame(speechPlayer_handle_t playerHandle, speechPlayer_frame_t* framePtr, int minFrameDuration, int fadeDuration, int userIndex, bool purgeQueue);
 int speechPlayer_synthesize(speechPlayer_handle_t playerHandle, int sampleCount, sample* sampleBuf); 
 int speechPlayer_getLastIndex(speechPlayer_handle_t playerHandle);
 void speechPlayer_terminate(speechPlayer_handle_t playerHandle);
